rag-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.59.4
sentence-transformers==3.3.1
chromadb==0.6.2
spacy==3.8.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
thefuzz==0.22.1
deep-translator==1.11.4
requests==2.32.3
pytz==2024.2